refactor: Rename packages to lowercase
Use this link to re-run the recipe: http://localhost:3000/recipes/org.openrewrite.staticanalysis.LowercasePackage
