--- demo-api/src/main/java/com/Example/Demo/Controller.java
@@ -1,16 +1,0 @@
-package	com.Example.Demo;
-
-import	lombok.extern.slf4j.Slf4j;
-import	org.springframework.web.bind.annotation.GetMapping;
-import	org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@Slf4j
-public	class	Controller	{
-
-	@GetMapping
-	public	Person	getPerson()	{
-		log.info("get	person!");
-		return	new	Person("fred");
-	}
-}
