--- demo-api/src/main/java/com/Example/Demo/DemoApplication.java
@@ -1,20 +1,0 @@
-package com.Example.Demo;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-
-@SpringBootApplication
-public class DemoApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DemoApplication.class, args);
-	}
-
-	@Bean("stupid")
-	public ObjectMapper myStupidObjectMapper() {
-		return new ObjectMapper().configure(SerializationFeature.INDENT_OUTPUT, true);
-	}
-}
