+++ demo-api/src/main/java/com/example/demo/Controller.java
@@ -1,0 +1,16 @@
+package com.example.demo;
+
+import	lombok.extern.slf4j.Slf4j;
+import	org.springframework.web.bind.annotation.GetMapping;
+import	org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@Slf4j
+public	class	Controller	{
+
+	@GetMapping
+	public	Person	getPerson()	{
+		log.info("get	person!");
+		return	new	Person("fred");
+	}
+}
