+++ demo-api/src/main/java/com/example/demo/DemoApplication.java
@@ -1,0 +1,20 @@
+package com.example.demo;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+
+@SpringBootApplication
+public class DemoApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DemoApplication.class, args);
+	}
+
+	@Bean("stupid")
+	public ObjectMapper myStupidObjectMapper() {
+		return new ObjectMapper().configure(SerializationFeature.INDENT_OUTPUT, true);
+	}
+}
